fix: quick refactor
