--- evals/transcription/src/core/ami/constants.py
@@ -1,2 +1,0 @@
-TARGET_SAMPLE_RATE = 16000
-STEREO_CHANNELS = 2
