+++ evals/transcription/src/core/ami/loader.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from datasets import load_dataset
 
 from . import audio, cache
-from .constants import TARGET_SAMPLE_RATE
+from ...constants import TARGET_SAMPLE_RATE
 from .metadata import load_or_build_metadata
 from .selection import MeetingSegment, select_segments
 from .types import Sample
